unify common request forwarder logic
